Implemented 3 helper functions for sorting.
is_negative(c) checks the final bit of a multi-bit encryption to see if it's a 1. After a subtraction, the most significant bit is the sign bit, so this will be 1 when the subtraction result is negative.

_if(cond, if_true, if_false) takes 3 parameters; the first is a single encrypted bit, the others are encrypted numbers (lists of encrypted bits). If the single encrypted bit is a 1, it will return something equivalent to if_true; otherwise it will return something equivalent to if_false.

min_max(c_1, c_2) takes two encrypted numbers (two lists of encrypted bits) and returns two encrypted numbers. The first return value is guaranteed to not be greater than the second return value (after decrypting).